refactor(profile): make profile stick to new Domain Structure
