+++ app/Domains/Admin/Database/Seeders/AdminUserSeeder.php
@@ -4,7 +4,7 @@ namespace App\Domains\Admin\Database\Seeders;
 
 use App\Domains\Auth\Models\User;
 use App\Domains\Auth\Models\Role;
-use App\Domains\Profile\Models\Profile;
+use App\Domains\Profile\Private\Models\Profile;
 use Illuminate\Database\Seeder;
 use Illuminate\Support\Facades\Hash;
