+++ app/Domains/Auth/Tests/Feature/RegisterTest.php
@@ -1,6 +1,6 @@
 <?php
 
-use App\Domains\Profile\Models\Profile;
+use App\Domains\Profile\Private\Models\Profile;
 use App\Domains\Auth\Events\UserRegistered;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Support\Facades\Auth;
