+++ app/Domains/Profile/Private/Listeners/ClearProfileCacheOnEmailVerified.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace App\Domains\Profile\Listeners;
+namespace App\Domains\Profile\Private\Listeners;
 
 use App\Domains\Auth\Events\EmailVerified;
-use App\Domains\Profile\Services\ProfileCacheService;
+use App\Domains\Profile\Private\Services\ProfileCacheService;
 use Illuminate\Contracts\Events\ShouldHandleEventsAfterCommit;
 
 class ClearProfileCacheOnEmailVerified implements ShouldHandleEventsAfterCommit
