+++ app/Domains/Profile/Private/Resources/lang/fr/events.php
