+++ bootstrap/providers.php
@@ -12,7 +12,7 @@ return [
     App\Domains\Home\Providers\HomeServiceProvider::class,
     App\Domains\Dashboard\Providers\DashboardServiceProvider::class,
     App\Domains\Auth\Providers\AuthServiceProvider::class,
-    App\Domains\Profile\PublicApi\Providers\ProfileServiceProvider::class,
+    App\Domains\Profile\Public\Providers\ProfileServiceProvider::class,
     App\Domains\StoryRef\Providers\StoryServiceRefProvider::class,
     App\Domains\Comment\PublicApi\Providers\CommentServiceProvider::class,
     App\Domains\Story\Providers\StoryServiceProvider::class,
